Move HpgRovProduct class to its own files.
